backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
asyncpg>=0.30.0
sqlalchemy[asyncio]>=2.0.36
alembic>=1.14.0
python-multipart>=0.0.18
python-jose[cryptography]>=3.4.0
passlib[bcrypt]>=1.7.4
pdfplumber>=0.11.0
python-docx>=1.1.0
readability-lxml>=0.8.1
httpx>=0.28.0
python-dotenv>=1.0.1
dashscope>=1.22.0
tenacity>=9.0.0
beautifulsoup4>=4.12.0
lxml>=5.3.0
celery[redis]>=5.4.0
redis>=5.2.0
aiofiles>=24.1.0
Pillow>=11.0.0
